Fix a thread blocking problem.
First use notifyAll instead of just notify to wakeup other threads. Also
use a timed wait as a workarounf of a (yet) unknown behaviour of
synhcronized/notifyAll: sometimes the waiting thread is not notified
and waits forever. Use a 100ms timed wait overcomes this problem.
TODO: check and clarify this behaviour.