Fix re-use of RawPackets.
Set length, offset and buffer data in re-used RawPacket objects. In RawPacket
change handling of length: only the length field holds the used and valid
data length. The length of the buffer array may be larger. Raw packet
implementation uses this to re-use existing buffer arrays as much as
possible.