Fix issues reported by Ingo on the dev mailing list where Jitsi fails to shutdown properly.
The fix relies on just calling System.exit() when we choose to shut down and the felix shutdown hooks take care of stopping everything nicely.
I've replaced the System.exit() call in the shutdown-timeout plugin with Runtime.getRuntime().halt() which doesn't run the shutdown hooks in felix so can't be blocked.

We'll need to keep a close eye on this to ensure that we're not resorting to the halt() in the shutdown-plugin in the mainline case. In future I hope to add a thread dump to the log just before we call .halt() to help diagnose problems in this area.