Add contact dialog: support prompts in text fields
This commit adds a function which makes it possible to add faint gray
prompt to a text field. The foreground color of the prompt's text is the
one of the text field, but half transparent. As soon as text is added to
the text field, the prompt will vanish and it will re-appear when the
text field is emptied again.