Optimized groupNode (sorted) add operations.
Original implementation used Collections.sort(...) to sort the elements
after having added a new element, which is typically a O(n log n)
operation, but O(n) in the best case which we are typically close to if
we insert a single element in an otherwise sorted vector. Afterwards
the index would be queried using getIndex(...) which is a O(n)
operation.

These operations are sufficiently slow that inserting a large number of
entries causes 100% CPU usage for a significant amount of time. In the
problematic case it would need to insert around 16000 IRC chat rooms
into this implementation, which caused very bad slow downs of Jitsi,
since it is the AWT thread that is doing the work.

The new implementation searching for the exact location at which to do a
sorted insertion, inserts the element at that location and returns that
index. It utilises a binarySearch, which is O(log n), to search for the
insertion point and since we then know the index, we do not have to look
that up.

In case of the (unsorted) add, since we know insertion behaviour we can
predict the insertion point and return that instead of doing an O(n)
getIndex(...) call for the just inserted element.

The largest deviation from the original implementation has to do with
the sorting operation. Now we insert in a way that a sorted list stays
sorted, however if other code would insert elements in an unsorted
manner into a sorted groupNode instance, then the new implementation
will not automatically correct itself, like the sort(...) operation
would.