Increased chat message buffer from 3000 to 50000.
Increased the buffer to ensure that the buffer is reasonable sufficient
for a reading back a decent conversation. The buffer value represents
the number of chars of "user data" stored in the HTML document. I
suspect things such as attribute data is also considered user data.
Therefore there is quite a "penalty" on keeping the original message
stored in an attribute, encoded in base64.