Fix for OTR UI in case of multiple OTR session.
Okay, so this was a real PITA. The otr4j library behaves as expected.
The OTR plugin, on the other hand, hides controls and does not get
updated on the actual selected resouce (chat transport) at all. So as a
consequence the OTR plugin is actually giving a lot of misinformation.

There is still an outstanding issue where you establish an OTR encrypted
session for a particular session ID and the OTR session selector box
does not automatically select the encrypted session. I'm not sure if
this should be the desired behaviour, though. What if there are 2 OTR
encrypted sessions, would you want to switch?

That is something to think about ...