Fixes the contacts 'call button' disappearing on multiresource contacts
This changes the behaviour from "last connected resource" or "highest
priority resource" to "most capable resource" regarding the display of
buttons next to this contact. We now check every resource the contact
has connected and we will consider the one with the most capabilities.
This elimates the very frustrating scenario in which the call
capabilites are disabled on a contact who connected additionally a
non-jingle client.

Note that there exists separate issue in which the contact logs off from
only one resource, but gets displayed as completely offline. In that
case the buttons are still displayed wrong, if the contact connects a
non-jingle client on afterwards.